requirements.txt
# Task 1 dependencies
emoji>=2.0.0

# Task 2 dependencies
python-dotenv>=1.0.0
pdfplumber>=0.10.0
langchain-groq>=0.1.0
langchain>=0.1.0
